Create error versions of T
In B5X, we had an issue where there were two options for Locale providers and one file used the wrong one, leading to a perplexing issue with partial translations. If the T function that initialized the empty context had errored instead of falling back to something halfway useful, the problem would have been a lot clearer.

Here I provide functions to make that kind of setup easy so that consumers don't need to maintain them.