Make glob a real dependency
We use glob in extract-strings, which is exported for use by consumers, so it should be an actual dependency, not a dev dependency. It could be a peer dependency, but there's no specific reason to require it to be synced with whatever version of `glob` a consumer may be using since it's just a dev tool.

While doing this I realized that `npm install` will change `package-lock.json` to use the same indentation style as `package.json`. I missed this when setting up Prettier before, so telling Prettier to ignore `package-lock.json` doesn't actually help. I could send `package.json` back to using spaces, but tabs are preferred for accessibility.

I've also run `npm pkg fix` here to address some warnings logged when publishing.